+++ Modules/GuildActions.lua
@@ -496,17 +496,21 @@ function GuildActions:CopyName(member)
 		return
 	end
 
+	local editBoxWidth = 300
 	StaticPopupDialogs["BFL_GUILD_COPY_NAME"] = {
 		text = T("GUILD_COPY_NAME_TITLE", "Copy Name"),
 		button1 = CLOSE or OKAY or "Close",
 		hasEditBox = true,
-		editBoxWidth = 300,
+		editBoxWidth = editBoxWidth,
 		OnShow = function(dialog)
 			local editBox = GetEditBox(dialog)
 			if editBox then
 				editBox:SetText(copyText)
 				editBox:HighlightText()
 				editBox:SetFocus()
+			end
+			if BFL.BrokerUtils and BFL.BrokerUtils.FixCopyStaticPopupLayout then
+				BFL.BrokerUtils.FixCopyStaticPopupLayout(dialog, editBoxWidth)
 			end
 		end,
 		EditBoxOnEnterPressed = function(editBox)
@@ -520,7 +524,10 @@ function GuildActions:CopyName(member)
 		hideOnEscape = true,
 		preferredIndex = 3,
 	}
-	StaticPopup_Show("BFL_GUILD_COPY_NAME")
+	local dialog = StaticPopup_Show("BFL_GUILD_COPY_NAME")
+	if dialog and BFL.BrokerUtils and BFL.BrokerUtils.FixCopyStaticPopupLayout then
+		BFL.BrokerUtils.FixCopyStaticPopupLayout(dialog, editBoxWidth)
+	end
 end
 
 function GuildActions:SetNicknameDialog(member)